Split keyboard code out into its own file
