+++ include/backend/libinput.h
@@ -5,6 +5,7 @@
 #include <wlr/common/list.h>
 #include <wayland-server-core.h>
 #include "backend/udev.h"
+#include "types.h"
 
 struct wlr_backend_state {
 	struct wlr_backend *backend;
@@ -21,8 +22,17 @@ struct wlr_backend_state {
 void wlr_libinput_event(struct wlr_backend_state *state,
 		struct libinput_event *event);
 
+struct wlr_input_device *get_appropriate_device(
+		enum wlr_input_device_type desired_type,
+		struct libinput_device *device);
+
 struct wlr_keyboard_state {
 	struct libinput_device *handle;
 };
+
+void handle_keyboard_key(struct libinput_event *event,
+		struct libinput_device *device);
+struct wlr_keyboard *wlr_libinput_keyboard_create(
+		struct libinput_device *device);
 
 #endif
